ZipService

+++ BookMarketRest/src/main/java/com/example/dto/requestDto/AuthorRequestDto.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class AuthorRequestDto {
     private String name;
-    private Long zipcodeId;
+    private Integer zipcodeId;
 }
